C++: Rewrite `reachable` without mutual recursion
This gives a measurable speedup, and I think it makes the code clearer.
Before this, the `reachable` and `successors_adapted` predicates were
mutually recursive together with four other predicates over two files.

This adds `successors_adapted` to the global name space, which is not
really desired but not harmful either.