C++: Speed up variableLiveOnEntryToBlock in IR
This predicate computed a local CP between all defs and uses of the same
virtual variable in a basic block. This wasn't a problem in
`unaliased_ssa`, but it became a huge problem in `aliased_ssa`, probably
because many variables can be modelled with a single virtual variable
there.

This commit reduces the evaluation time of
`variableLiveOnEntryToBlock#ff#antijoin_rhs` on Wireshark from 80
_minutes_ to a few seconds.