C++: IR query for redundant null check
This new query is not written because it's the most interesting query we
could write but because it's an IR-based query whose results are easy to
verify.

This query is not ready for LGTM, but it can go in the standard suites
so we can observe its performance in nightly jobs.