Import CppType through a new module signature
The `TypeDomain` types had to move from `IRCppLanguage` to `CppType`. It
was a breach of abstraction that they lived in IRCppLanguage because
they were actually used in `CppType`. The only reason it worked before
was that there happened to be a type of the same name coming in via
`import cpp`, and those types happened to be compatible.